Removed circular dependencies; migrated GeoImageTMS static functions to lib/common
