+++ include/vapor/ImageParams.h
@@ -3,7 +3,7 @@
 
 #include <vapor/RenderParams.h>
 #include <vapor/DataMgr.h>
-#include <vapor/GeoImageTMS.h>
+#include <vapor/TMSUtils.h>
 
 namespace VAPoR 
 {
@@ -40,8 +40,8 @@ public:
   {
     BeginGroup("Set image path");
     SetValueString( _fileNameTag, "Set image file path", file );
-    if ( GeoImageTMS::IsTMSFile( file ) ) {
-      int numTMSLODs = GeoImageTMS::GetNumTMSLODs( file );
+    if ( Wasp::TMSUtils::IsTMSFile( file ) ) {
+      int numTMSLODs = Wasp::TMSUtils::GetNumTMSLODs( file );
       _setNumTMSLODs( numTMSLODs );
     }
     EndGroup();
