+++ lib/render/GeoImageTMS.cpp
@@ -14,6 +14,7 @@
 #include <vapor/Proj4API.h>
 #include <vapor/GeoUtil.h>
 #include <vapor/GeoTileMercator.h>
+#include <vapor/TMSUtils.h>
 #include <vapor/GeoImageTMS.h>
 
 using namespace VAPoR;
@@ -53,56 +54,6 @@ GeoImageTMS::~GeoImageTMS()
 	if (_geotile) delete _geotile;
 }
 
-bool GeoImageTMS::IsTMSFile( std::string path ) {
-    if ( path.rfind(".tms", path.size()-4) != string::npos ) {
-        return true;
-    }
-    return false;
-}
-
-std::string GeoImageTMS::TilePath(
-    string file, size_t tileX, size_t tileY, int lod
-) {
-    // If we're given a file instead of a directory, remove the .tms extension
-    //
-    if ( file.rfind(".tms", file.size()-4) != string::npos ) {
-        file.erase( file.length()-4, 4 );
-    }
-
-    size_t tmsTileY = tileY;
-
-    ostringstream oss;
-    oss << file;
-    oss << "/";
-    oss << lod;
-    oss << "/";
-    oss << tileX;
-    oss << "/";
-    oss << tmsTileY;
-
-    string base = oss.str();
-
-    string path = base + ".tif";
-
-    struct stat statbuf;
-    if (stat(path.c_str(), &statbuf) == 0)  return(path);
-
-    path = base + ".tiff";
-
-    if (stat(path.c_str(), &statbuf) == 0) return (path);
-
-    // Tile does not exist
-    //
-    return("");
-}
-
-int GeoImageTMS::GetNumTMSLODs( std::string file ) {
-    int lod = 0;
-    while ( TilePath( file, 0, 0, lod ) != "" ) {
-        lod++;
-    }
-    return lod;
-}
 
 int GeoImageTMS::Initialize(string dir, vector <double> times) {
 
@@ -115,7 +66,7 @@ int GeoImageTMS::Initialize(string dir, vector <double> times) {
 
 	// Find the maximum available LOD in the TMS database. 
 	//
-    int lod = GetNumTMSLODs( _dir ) - 1;
+    int lod = TMSUtils::GetNumTMSLODs( _dir ) - 1;
 	if (lod<0) {
 		SetErrMsg("Failed to initialize TMS directory %s", _dir.c_str());
 		return(-1);
@@ -317,7 +268,7 @@ int GeoImageTMS::_tileSize(
 	string dir, size_t tileX, size_t tileY, int lod, size_t &w, size_t &h
 ) {
 	
-	string path = TilePath(dir, tileX, tileY, lod);
+	string path = TMSUtils::TilePath(dir, tileX, tileY, lod);
 	if (path.empty()) {
 		SetErrMsg("Tile %d %d %d does not exist", tileX, tileY, lod);
 		return(-1);
@@ -347,7 +298,7 @@ int GeoImageTMS::_tileRead(
 		"GeoImageTMS::_tileRead(%s,%d,%d,%d)", dir.c_str(), tileX, tileY, lod
 	);
 
-	string path = TilePath(dir, tileX, tileY, lod);
+	string path = TMSUtils::TilePath(dir, tileX, tileY, lod);
 	if (path.empty()) {
 		SetErrMsg("Tile %d %d %d does not exist", tileX, tileY, lod);
 		return(-1);
